Split out testing methods into a separate trait.
